answer_generation_no_retrieval/pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "answer-generation-no-retrieval"
version = "0.1.0"
description = "Closed-book ablation: answer ViDoRe v3 queries without retrieval"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.5.0",
    "httpx>=0.27.0",
    "loguru>=0.7.3",
    "pandas>=2.0.0",
    "python-dotenv>=1.2.1",
    "tqdm",
    "typer",
    "visidata>=3.3",
]

[tool.ruff]
line-length = 99
